Rename api entities to just entities (route level)
