+++ src/domain/chains/entities/__tests__/rpc-uri.factory.ts
@@ -1,6 +1,6 @@
-import { RpcUri } from '../../../../routes/chains/openapi/api-rpc-uri';
 import { faker } from '@faker-js/faker';
 import { RpcUriAuthentication } from '../rpc-uri-authentication.entity';
+import { RpcUri } from '../rpc-uri.entity';
 
 export default function (
   authentication?: RpcUriAuthentication,
