+++ src/routes/chains/chains.controller.ts
@@ -1,5 +1,4 @@
 import { Controller, Get, Param } from '@nestjs/common';
-import { Page } from '../../common/entities/page.entity';
 import { ChainsService } from './chains.service';
 import { PaginationData } from '../../common/pagination/pagination.data';
 import { RouteUrlDecorator } from '../../common/decorators/route.url.decorator';
@@ -7,8 +6,9 @@ import { PaginationDataDecorator } from '../../common/decorators/pagination.data
 import { Backbone } from '../../domain/backbone/entities/backbone.entity';
 import { Chain } from '../../domain/chains/entities/chain.entity';
 import { ApiOkResponse, ApiTags } from '@nestjs/swagger';
-import { Backbone as ApiBackbone } from './openapi/api-backbone';
-import { Page as ApiPage } from './openapi/api-chain-page';
+import { Backbone as ApiBackbone } from './entities/backbone.entity';
+import { Page } from '../../domain/entities/page.entity';
+import { ChainPage } from './entities/chain-page.entity';
 
 @ApiTags('chains')
 @Controller({
@@ -18,7 +18,7 @@ import { Page as ApiPage } from './openapi/api-chain-page';
 export class ChainsController {
   constructor(private readonly chainsService: ChainsService) {}
 
-  @ApiOkResponse({ type: ApiPage })
+  @ApiOkResponse({ type: ChainPage })
   @Get()
   async getChains(
     @RouteUrlDecorator() routeUrl: URL,
