+++ src/routes/chains/entities/block-explorer-uri-template.entity.ts
