+++ src/routes/chains/entities/chain-page.entity.ts
@@ -1,0 +1,8 @@
+import { Page } from '../../../common/entities/page.entity';
+import { Chain } from './chain.entity';
+import { ApiProperty } from '@nestjs/swagger';
+
+export class ChainPage extends Page<Chain> {
+  @ApiProperty({ type: Chain })
+  results: Chain[];
+}
