+++ src/routes/chains/entities/gas-price-fixed.entity.ts
